+++ alymova.ksenia/T3/shapes.hpp
@@ -5,6 +5,10 @@
 
 namespace alymova
 {
+  struct DelimiterIO
+  {
+    char exp;
+  };
   struct Point
   {
     int x, y;
@@ -13,18 +17,11 @@ namespace alymova
   {
     std::vector< Point > points;
   };
-  struct DelimiterIO
-  {
-    char exp;
-  };
 
+  std::istream& operator>>(std::istream& in, DelimiterIO&& object);
   std::istream& operator>>(std::istream& in, Point& point);
   std::istream& operator>>(std::istream& in, Polygon& polygon);
-  std::istream& operator>>(std::istream& in, DelimiterIO&& object);
   std::ostream& operator<<(std::ostream& out, const Point& point);
   std::ostream& operator<<(std::ostream& out, const Polygon& polygon);
-
-  double multPoints(const Point& point1, const Point& point2);
-  bool isPolygonEven(const Polygon& polygon);
 }
 #endif
